feat: add is active styles
